bugs fixed, manual music input missing
